refactor: RacingCarGame 구조 변경 (step3)
