+++ src/main/java/step3/README.md
@@ -1,6 +1,9 @@
 ## 기능 요구사항
-1. 자동차수와 시도 횟수에 대한 입력값을 받는다.
-2. 자동차 별 0에서 9사이의 random 값을 구한다.
-3. random 값이 0미만 9초과의 수일 경우, RuntimeException 이 발생한다.
-4. random 값이 4 이상일 경우에만, 해당 자동차는 전진한다.(- 를 출력한다.) : **시도 횟수만큼 반복한다.**
-5. random 값이 4 미만일 경우, 자동차는 정지한다.
+1. 자동차 수를 입력 받는다.
+2. 라운드 수를 입력 받는다.
+3. 입력 값만큼 차를 생성한다.
+4. 랜덤 값을 생성한다.
+5. 랜덤 값이 0에서 9 사이의 값이 아닐 경우, 에러가 발생한다.
+6. 랜덤 값이 전진 조건(4 이상)에 해당하는지 확인한다.
+7. 자동차는 전진 조건에 따라 전진 또는 정지한다.
+8. 화면에 결과 값을 출력한다.
