--- src/main/java/step3/RacingCarGame.java
@@ -1,8 +1,0 @@
-package step3;
-
-public class RacingCarGame {
-    public static void main(String[] args) {
-        CarRacing racing = new CarRacing(new RandomNumMovingStrategy());
-        racing.startGame();
-    }
-}
