+++ src/main/java/step3/RacingCarGameApplication.java
@@ -1,0 +1,11 @@
+package step3;
+
+import step3.controller.RacingCarController;
+import step3.strategy.RandomNumMovingStrategy;
+
+public class RacingCarGameApplication {
+    public static void main(String[] args) {
+        RacingCarController racing = new RacingCarController(new RandomNumMovingStrategy());
+        racing.startGame();
+    }
+}
