--- src/main/java/step3/RacingProperty.java
@@ -1,6 +1,0 @@
-package step3;
-
-public enum RacingProperty {
-    CAR_COUNT,
-    ATTEMPT_COUNT
-}
