+++ src/main/java/step3/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.strategy;
 
 public interface MovingStrategy {
     boolean isForward();
