+++ src/main/java/step3/strategy/RandomNumMovingStrategy.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.strategy;
 
 import step3.util.RandomNumberGenerator;
