+++ src/main/java/step3/view/ResultView.java
@@ -1,4 +1,4 @@
-package step3.util;
+package step3.view;
 
 public class ResultView {
     private static final String MOVE_SYMBOL = "-";
