+++ src/test/java/step3/CarTest.java
@@ -2,6 +2,8 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Car;
+import step3.strategy.MovingStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
